CityPopularity {googleVis}	R Documentation

CityPopularity: googleVis example data set

Description

Example data set to illustrate the use of the googleVis package.

Usage

data(CityPopularity)

Format

A data frame with 6 observations on the following 2 variables.

City

a factor with levels Boston Chicago Houston Los Angeles Miami New York

Popularity

a numeric vector

Source

Google Geo Map API: https://google-developers.appspot.com/chart/interactive/docs/gallery/geomap.html

Examples

data(CityPopularity)
str(CityPopularity)

[Package googleVis version 0.6.4 Index]