OpenClose: googleVis example data set

Example data set to illustrate the use of the googleVis package.

Usage

data(OpenClose)

Format

A data frame with 5 observations on the following 5 variables.

Weekday

a factor with levels Fri Mon Thurs Tues Wed

Low

a numeric vector

Open

a numeric vector

Close

a numeric vector

High

a numeric vector

Source

Google Visualisation: Candlestick Chart https://developers.google.com/chart/interactive/docs/gallery/candlestickchart?csw=1

Examples

OpenClose
#>   Weekday Low Open Close High
#> 1     Mon  20   28    38   45
#> 2    Tues  31   38    55   66
#> 3     Wed  50   55    77   80
#> 4   Thurs  50   77    66   77
#> 5     Fri  15   66    22   68
plot(gvisCandlestickChart(OpenClose, options=list(legend='none')))