R Interface to Google Charts

R interface to Google Charts API, allowing users to create interactive charts based on data frames. Charts are displayed locally via the R HTTP help server. A modern browser with Internet connection is required and for some charts Flash. The data remains local and is not uploaded to Google.

Please visit the project web site for more information: https://github.com/mages/googleVis.

You find further notes on Markus' blog: https://magesblog.com/tags/googlevis/

Author

Markus Gesmann, Diego de Castillo

Note

See vignette("googleVis") for an introduction to the googleVis package.

References

Google Charts API: https://developers.google.com/chart/

Google Terms of Use: https://developers.google.com/terms/

Google Maps API Terms of Service: https://cloud.google.com/maps-platform/terms/

Examples

if (FALSE) { # \dontrun{
  demo(googleVis)
  ## For other demos see
  demo(package='googleVis')  
  } # }