Google Histogram Chart with R histogram gvisHistogram

The gvisHistogram function reads a data.frame and creates text output referring to the Google Visualisation API, which can be included into a web page, or as a stand-alone page. The actual chart is rendered by the web browser using SVG or VML.

Usage

gvisHistogram(data, options = list(), chartid)

Arguments

data

a data.frame to be displayed as a histogram. Each column will be displayed as a histogram.

options

list of configuration options, see

https://developers.google.com/chart/interactive/docs/gallery/histogram#Configuration_Options

The parameters can be set via a named list. The parameters have to map those of the Google documentation.

Boolean arguments are set to either TRUE or FALSE, using the R syntax.

Google API parameters with a single value and with names that don't include a "." are set like one would do in R, that is options=list(width=200, height=300). Exceptions to this rule are the width and height options for gvisAnnotatedTimeLine and gvisAnnotationChart. For those two functions, width and height must be character strings of the format "Xpx", where X is a number, or "automatic". For example, options=list(width="200px", height="300px").

Google API parameters with names that don't include a ".", but require multivalues are set as a character, wrapped in "[ ]" and separated by commas, e.g. options=list(colors="['#cbb69d', '#603913', '#c69c6e']")

Google API parameters with names that do include a "." present parameters with several sub-options and have to be set as a character wrapped in " ". The values of those sub-options are set via parameter:value. Boolean values have to be stated as 'true' or 'false'. For example the Google documentaion states the formating options for the vertical axis and states the parameter as vAxis.format. Then this paramter can be set in R as:options=list(vAxis="{format:'#,###%'}").

If several sub-options have to be set, e.g. titleTextStyle.color, titleTextStyle.fontName and titleTextStyle.fontSize, then those can be combined in one list item such as:options=list(titleTextStyle="{color:'red', fontName:'Courier', fontSize:16}")

paramters that can have more than one value per sub-options are wrapped in "[ ]". For example to set the labels for left and right axes use:options=list(vAxes="[{title:'val1'}, {title:'val2'}]")

gvis.editor a character label for an on-page button that opens an in-page dialog box enabling users to edit, change and customise the chart. By default no value is given and therefore no button is displayed.

For more details see the Google API documentation and the R examples below.

chartid

character. If missing (default) a random chart id will be generated based on chart type and tempfile.

Value

gvisHistogram returns list of class"gvis" and "list". An object of class "gvis" is a list containing at least the following components:

type

Google visualisation type

chartid

character id of the chart object. Unique chart ids are required to place several charts on the same page.

html

a list with the building blocks for a page

header

a character string of a html page header: <html>...<body>,

chart

a named character vector of the chart's building blocks:

jsHeader

Opening <script> tag and reference to Google's JavaScript library.

jsData

JavaScript function defining the input data as a JSON object.

jsDrawChart

JavaScript function combing the data with the visualisation API and user options.

jsDisplayChart

JavaScript function calling the handler to display the chart.

jsFooter

End tag </script>.

jsChart

Call of the jsDisplayChart function.

divChart

<div> container to embed the chart into the page.

caption

character string of a standard caption, including data name and chart id.

footer

character string of a html page footer: </body>...</html>, including the used R and googleVis version and link to Google's Terms of Use.

References

Google Chart Tools API: https://developers.google.com/chart/interactive/docs/gallery/histogram

See also

Author

Markus Gesmann markus.gesmann@gmail.com,

Diego de Castillo decastillo@gmail.com

Examples

## Please note that by default the googleVis plot command
## will open a browser window and requires an internet
## connection to display the visualisation.

hist1 <- gvisHistogram(dino)
plot(hist1)

## Histogram of the top 20 countries
pop <- Population[1:20,c("Country", "Population")]
pop=transform(pop, Population=round(Population/1e6))

hist2 <- gvisHistogram(pop, option=list(title="Country Populations",
                                    legend="{ position: 'none' }",
                                    colors="['green']"))
plot(hist2)
                                    
set.seed(123)
dat=data.frame(A=rpois(100, 20), 
               B=rpois(100, 5), 
               C=rpois(100, 50))
hist3 <- gvisHistogram(dat, options=list(
                       legend="{ position: 'top', maxLines: 2 }",
                       colors="['#5C3292', '#1A8763', '#871B47']"))

plot(hist3)