renderGvis

This function lets you use googleVis charts as Shiny output. Use it to wrap a googleVis-generating function that you assign to an output element in server.R; then create an htmlOutput with the same name in ui.R.

Usage

renderGvis(expr, env = parent.frame(), quoted = FALSE)

Arguments

expr

An expression that returns a gvis object.

env

The environment in which to evaluate expr

quoted

Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

Value

Returns a function that can be assigned to a Shiny output element.

Details

Use a googleVis Chart as Shiny Output

Details

More information about shiny is available online: https://shiny.posit.co/. You find further examples with googleVis on shiny on mages' blog: https://magesblog.com/tags/shiny/

Author

Joe Cheng, joe@rstudio.com

Examples

  
if (FALSE) { # \dontrun{
# To run this example:
shiny::runApp(system.file("shiny/", package="googleVis"))
# server.R
library(googleVis)

shinyServer(function(input, output) {
  datasetInput <- reactive({
    switch(input$dataset,
           "rock" = rock,
           "pressure" = pressure,
           "cars" = cars)
  })
  
  output$view <- renderGvis({
    gvisScatterChart(datasetInput(),
                     options=list(title=paste('Data:',input$dataset)))
  })
})

# ui.R
shinyUI(pageWithSidebar(
  headerPanel("googleVis on Shiny"),
  sidebarPanel(
    selectInput("dataset", "Choose a dataset:",
                choices = c("rock", "pressure", "cars"))
  ),
  mainPanel(
    htmlOutput("view")
  )
))
} # }