googleVis-package {googleVis}	R Documentation

R Interface to Google Charts

Description

R interface to Google Charts API, allowing users to create interactive charts based on data frames. Charts are displayed locally via the R HTTP help server. A modern browser with Internet connection is required and for some charts Flash. The data remains local and is not uploaded to Google.

Please visit the project web site for more information: https://github.com/mages/googleVis.

You find further notes on Markus' blog: https://magesblog.com/tags/googlevis/

Note

See vignette("googleVis") for an introduction to the googleVis package.

Author(s)

Markus Gesmann, Diego de Castillo

References

Google Charts API: https://developers.google.com/chart/

Google Terms of Use: https://developers.google.com/terms/

Google Maps API Terms of Service: https://developers.google.com/maps/terms

Examples

## Not run: 
  demo(googleVis)
  ## For other demos see
  demo(package='googleVis')	
  
## End(Not run)

[Package googleVis version 0.6.4 Index]